+++ src/config/mod.rs
@@ -91,6 +91,7 @@ pub struct Config {
     pub auto_open_downloads: bool,
     pub extra_yt_dlp_args: Vec<String>,
     pub duration_format: DurationFormat,
+    pub on_exit: Option<Arc<Vec<String>>>,
 }
 
 impl Default for Config {
@@ -158,6 +159,7 @@ pub struct ConfigFile {
     pub extra_yt_dlp_args: Vec<String>,
     pub auto_open_downloads: bool,
     pub duration_format: String,
+    pub on_exit: Option<Vec<String>>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Copy, Clone, PartialEq, Eq)]
@@ -228,6 +230,7 @@ impl Default for ConfigFile {
             extra_yt_dlp_args: Vec::new(),
             auto_open_downloads: true,
             duration_format: "%m:%S".to_string(),
+            on_exit: None,
         }
     }
 }
@@ -376,6 +379,9 @@ impl ConfigFile {
             extra_yt_dlp_args: self.extra_yt_dlp_args,
             auto_open_downloads: self.auto_open_downloads,
             duration_format: DurationFormat::parse(&self.duration_format)?,
+            on_exit: self.on_exit.map(|arr| {
+                Arc::new(arr.into_iter().map(|v| tilde_expand(&v).into_owned()).collect_vec())
+            }),
         };
 
         if skip_album_art_check {